Get newSession function working nicely
